# Speed up bothify by making it via one run