results of measurements (almost twice for bothify)
before
```
Benchmark                            Mode  Cnt      Score       Error   Units
DatafakerTemplateStrings.bothify    thrpt   10  15713.850 ±  2229.838  ops/ms
DatafakerTemplateStrings.letterify  thrpt   10  42921.869 ±  4525.499  ops/ms
DatafakerTemplateStrings.numerify   thrpt   10  34054.217 ± 10869.288  ops/ms
```
after
```
Benchmark                            Mode  Cnt      Score      Error   Units
DatafakerTemplateStrings.bothify    thrpt   10  26830.873 ±  849.909  ops/ms
DatafakerTemplateStrings.letterify  thrpt   10  43706.715 ± 5250.235  ops/ms
DatafakerTemplateStrings.numerify   thrpt   10  39695.582 ± 9124.170  ops/ms
```